+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxWits.hs
@@ -216,9 +216,9 @@ pattern ShelleyTxWits {addrWits, scriptWits, bootWits} <-
 
 shelleyEqTxWitsRaw :: EraTxWits era => TxWits era -> TxWits era -> Bool
 shelleyEqTxWitsRaw txWits1 txWits2 =
-  liftEq eqRaw (txWits1 ^. addrTxWitsL) (txWits2 ^. addrTxWitsL)
+  txWits1 ^. addrTxWitsL == txWits2 ^. addrTxWitsL
     && liftEq eqRaw (txWits1 ^. scriptTxWitsL) (txWits2 ^. scriptTxWitsL)
-    && liftEq eqRaw (txWits1 ^. bootAddrTxWitsL) (txWits2 ^. bootAddrTxWitsL)
+    && txWits1 ^. bootAddrTxWitsL == txWits2 ^. bootAddrTxWitsL
 
 instance EraScript era => DecCBOR (Annotator (ShelleyTxWitsRaw era)) where
   decCBOR = decodeWits