+++ libs/cardano-ledger-core/test/Test/Cardano/Ledger/AddressSpec.hs
@@ -57,7 +57,7 @@ spec =
             BootstrapWitness
               { bwKey = shelleyVKey
               , bwChainCode = chainCode
-              , bwSig = sig
+              , bwSignature = sig
               , bwAttributes = serialize' byronProtVer $ Byron.addrAttributes byronAddr
               }
       pure $